fix bug for different os, cleanup
